# docs/api/openapi.current.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ink Buddy Chat Bot",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "[Scaffold] login",
        "description": "Returns 501 for valid requests. Success schema is a proposed contract, not implemented behavior.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "501": {
            "description": "Route scaffold only; business logic not implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-implementation-status": "scaffold"
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "[Scaffold] refresh token",
        "description": "Returns 501 for valid requests. Success schema is a proposed contract, not implemented behavior.",
        "operationId": "refresh_token_api_v1_auth_refresh_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "501": {
            "description": "Route scaffold only; business logic not implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-implementation-status": "scaffold"
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "[Scaffold] logout",
        "description": "Returns 501 for valid requests. Success schema is a proposed contract, not implemented behavior.",
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "501": {
            "description": "Route scaffold only; business logic not implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-implementation-status": "scaffold"
      }
    },
    "/api/v1/auth/guest-sessions": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "[Scaffold] create guest session",
        "description": "Returns 501 for valid requests. Success schema is a proposed contract, not implemented behavior.",
        "operationId": "create_guest_session_api_v1_auth_guest_sessions_post",
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestSessionResponse"
                }
              }
            }
          },
          "501": {
            "description": "Route scaffold only; business logic not implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-implementation-status": "scaffold"
      }
    },
    "/api/v1/auth/guest-sessions/current": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "[Scaffold] get guest session",
        "description": "Returns 501 for valid requests. Success schema is a proposed contract, not implemented behavior.",
        "operationId": "get_guest_session_api_v1_auth_guest_sessions_current_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestSessionResponse"
                }
              }
            }
          },
          "501": {
            "description": "Route scaffold only; business logic not implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-implementation-status": "scaffold"
      }
    },
    "/api/v1/auth/guest-sessions/current/claim": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "[Scaffold] claim guest session",
        "description": "Returns 501 for valid requests. Success schema is a proposed contract, not implemented behavior.",
        "operationId": "claim_guest_session_api_v1_auth_guest_sessions_current_claim_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestClaimResponse"
                }
              }
            }
          },
          "501": {
            "description": "Route scaffold only; business logic not implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-implementation-status": "scaffold"
      }
    },
...
    "/api/v1/images": {
      "post": {
        "tags": [
          "images"
        ],
        "summary": "Upload Image",
        "description": "Attaching a photo does not identify a product yet; call /product-search/by-image next.",
        "operationId": "upload_image_api_v1_images_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_image_api_v1_images_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageUploadResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/images/{image_id}/analysis": {
      "post": {
        "tags": [
          "images"
        ],
        "summary": "Analyze Image",
        "description": "Describe the photo and its visible attributes. An interpretation, not a confirmed product.\n\nThe first call runs the vision model (slow); later calls return the stored result.",
        "operationId": "analyze_image_api_v1_images__image_id__analysis_post",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Image Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageAnalysisResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/v1/images/{image_id}/ocr": {
      "post": {
        "tags": [
          "images"
        ],
        "summary": "Ocr Image",
        "description": "Read the text printed on the item. Approximate: verify before citing a model or SKU.\n\nThe first call runs the vision model (slow); later calls return the stored result.",
        "operationId": "ocr_image_api_v1_images__image_id__ocr_post",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Image Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageOcrResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/v1/images/{image_id}": {
      "get": {
        "tags": [
          "images"
        ],
        "summary": "[Scaffold] get image",
        "description": "Returns 501 for valid requests. Success schema is a proposed contract, not implemented behavior.",
        "operationId": "get_image_api_v1_images__image_id__get",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Image Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageDetailResponse"
                }
              }
            }
          },
          "501": {
            "description": "Route scaffold only; business logic not implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-implementation-status": "scaffold"
      },
      "delete": {
        "tags": [
          "images"
        ],
        "summary": "[Scaffold] delete image",
        "description": "Returns 501 for valid requests. Success schema is a proposed contract, not implemented behavior.",
        "operationId": "delete_image_api_v1_images__image_id__delete",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Image Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "501": {
            "description": "Route scaffold only; business logic not implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-implementation-status": "scaffold"
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_upload_image_api_v1_images_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "JPEG / PNG / WEBP, checked by content"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_image_api_v1_images_post"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "For OpenAPI ``responses=`` declarations."
      },
      "GuestClaimResponse": {
        "properties": {
          "guest_session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Guest Session Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "chat_sessions_claimed": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Chat Sessions Claimed"
          },
          "images_claimed": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Images Claimed"
          }
        },
        "type": "object",
        "required": [
          "guest_session_id",
          "user_id",
          "chat_sessions_claimed",
          "images_claimed"
        ],
        "title": "GuestClaimResponse"
      },
      "GuestSessionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "image_upload_limit": {
            "type": "integer",
            "const": 3,
            "title": "Image Upload Limit",
            "default": 3
          },
          "image_uploads_used": {
            "type": "integer",
            "maximum": 3.0,
            "minimum": 0.0,
            "title": "Image Uploads Used"
          },
          "image_uploads_remaining": {
            "type": "integer",
            "maximum": 3.0,
            "minimum": 0.0,
            "title": "Image Uploads Remaining"
          }
        },
        "type": "object",
        "required": [
          "id",
          "expires_at",
          "image_uploads_used",
          "image_uploads_remaining"
        ],
        "title": "GuestSessionResponse"
      },
...
      "ImageAnalysisResponse": {
        "properties": {
          "image_id": {
            "type": "string",
            "format": "uuid",
            "title": "Image Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "attributes": {
            "$ref": "#/components/schemas/ImageAttributes"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached",
            "description": "true = stored result from an earlier analysis of this image"
          }
        },
        "type": "object",
        "required": [
          "image_id",
          "description",
          "attributes",
          "cached"
        ],
        "title": "ImageAnalysisResponse",
        "description": "An interpretation of the photo, not a confirmed product."
      },
      "ImageAttributes": {
        "properties": {
          "is_stationery": {
            "type": "boolean",
            "title": "Is Stationery"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "description": "closest catalog category, or 'other'"
          },
          "brand_text": {
            "type": "string",
            "title": "Brand Text",
            "description": "brand text read off the item; empty if none"
          },
          "model_text": {
            "type": "string",
            "title": "Model Text",
            "description": "model/series text read off the item; empty if none"
          },
          "colors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Colors",
            "default": []
          },
          "features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Features",
            "description": "visible features: tip size, pack count, shape, material",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "is_stationery",
          "category",
          "brand_text",
          "model_text"
        ],
        "title": "ImageAttributes"
      },
      "ImageDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "ready | processing | failed"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "mime_type",
          "size_bytes",
          "created_at"
        ],
        "title": "ImageDetailResponse"
      },
      "ImageOcrResponse": {
        "properties": {
          "image_id": {
            "type": "string",
            "format": "uuid",
            "title": "Image Id"
          },
          "text": {
            "type": "string",
            "title": "Text",
            "description": "all segments joined by newlines; empty when no text was found"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/OcrSegmentOut"
            },
            "type": "array",
            "title": "Segments"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached"
          }
        },
        "type": "object",
        "required": [
          "image_id",
          "text",
          "segments",
          "cached"
        ],
        "title": "ImageOcrResponse",
        "description": "Approximate; verify before citing a model or SKU."
      },
      "ImageUploadResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "ready | processing | failed"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status"
        ],
        "title": "ImageUploadResponse"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 320,
            "minLength": 3,
            "title": "Email",
            "description": "Email format validation to be added with auth implementation"
          },
          "password": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "OcrSegmentOut": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "OcrSegmentOut"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "const": "bearer",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Expires In"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expires_in"
        ],
        "title": "TokenResponse"
      },
...
    }
  }
}